Fix cells flow
