+++ ExampleiOS/Source/Classes/CoreFoundation/Assets/Assets.swift
@@ -47,8 +47,8 @@ public enum Asset {
     public static let warning = ImageAsset(name: "Card/warning")
   }
   public enum Chat {
-    public static let comment = ImageAsset(name: "Chat/Comment")
-    public static let photo = ImageAsset(name: "Chat/Photo")
+    public static let comment = ImageAsset(name: "Chat/comment")
+    public static let photo = ImageAsset(name: "Chat/photo")
   }
   public enum Currency {
     public static let britain = ImageAsset(name: "Currency/Britain")
