+++ ExampleiOS/Source/Classes/UserStrories/Main/SwiftUI/Cells/BaseCell/CenterElements/CenterElementsSwiftUIView.swift
@@ -1,20 +1,17 @@
 //
-//  ImageLableComponentsSwiftUIView.swift
+//  LeadingComponentsSwiftUIView.swift
 //  ExampleiOS
 //
-//  Created on 22.06.2021.
+//  Created on 07.06.2021.
 //
 
 import SwiftUI
 import AdmiralTheme
+import AdmiralUIResources
 import AdmiralSwiftUI
 
 @available(iOS 14.0.0, *)
-struct ImageLableComponentsSwiftUIView: View {
-    
-    enum Constants {
-        static let tabFormatAccessibilityValue = "%@. Page %i of %i"
-    }
+struct CenterElementsSwiftUIView: View {
     
     @State private var selectedIndex: Int?
     @State private var isEnabledControlsState: Int = 0
@@ -22,77 +19,88 @@ struct ImageLableComponentsSwiftUIView: View {
     
     public var body: some View {
         let scheme = schemeProvider.scheme
-        NavigationContentView(navigationTitle: "Leading elements") {
+        NavigationContentView(navigationTitle: "Center elements") {
             scheme.backgroundColor.swiftUIColor
                 .edgesIgnoringSafeArea(.all)
-            VStack(alignment: .leading, spacing: 0.0) {
+            ScrollView(showsIndicators: false) {
                 StandardTab(items: [
                     StandartTabItem(text: "Default", accessibilityId: "SegmentControlDefault"),
                     StandartTabItem(text: "Disabled", accessibilityId: "SegmentControlDisabled")
                 ],
                             selection: $isEnabledControlsState)
                     .padding()
-                ScrollView(showsIndicators: false) {
+                LazyVStack(alignment: .leading) {
                     ListCell(
-                        leadingView: { LeadingCardListView(cardImage: Image(uiImage: Asset.Card.visa.image)) },
-                        centerView: { TitleListView(title: "Card Place") },
+                        centerView: { TitleListView(title: "Title") },
                         trailingView: { ArrowListView() },
                         isSelected:
                             Binding(
                                 get: { self.selectedIndex == 0 },
                                 set: { _, _ in self.selectedIndex = self.selectedIndex == 0 ? nil : 0 }
                             ))
-                        .accessibilityElement()
-                        .accessibility(identifier: "LeadingCardTitleArrowListCell")
                         .disabled(isEnabledControlsState != 0)
+                        .accessibilityElement()
+                        .accessibility(identifier: "TitleListView")
                     ListCell(
-                        leadingView: { ImageListView(image: Image(uiImage: Asset.Card.rnb.image)) },
-                        centerView: { TitleListView(title: "Lable Place") },
+                        centerView: { TitleSubtitleListView(title: "Title", subtitle: "Subtitle") },
                         trailingView: { ArrowListView() },
                         isSelected:
                             Binding(
                                 get: { self.selectedIndex == 1 },
                                 set: { _, _ in self.selectedIndex = self.selectedIndex == 1 ? nil : 1 }
                             ))
-                        .accessibilityElement()
-                        .accessibility(identifier: "FirstImageTitleArrowListCell")
                         .disabled(isEnabledControlsState != 0)
+                        .accessibilityElement()
+                        .accessibility(identifier: "TitleSubtitleListView")
                     ListCell(
-                        leadingView: { ImageNameListView(text: "IN") },
-                        centerView: { TitleListView(title: "Icon Name") },
+                        centerView: { SubtitleTitleListView(title: "Title", subtitle: "Subtitle") },
                         trailingView: { ArrowListView() },
                         isSelected:
                             Binding(
                                 get: { self.selectedIndex == 2 },
                                 set: { _, _ in self.selectedIndex = self.selectedIndex == 2 ? nil : 2 }
                             ))
-                        .accessibilityElement()
-                        .accessibility(identifier: "SecondImageTitleArrowListCell")
                         .disabled(isEnabledControlsState != 0)
+                        .accessibilityElement()
+                        .accessibility(identifier: "SubtitleTitleListView")
                     ListCell(
-                        leadingView: { ImageBackgroundListView(image: Image(uiImage: Asset.Main.gem.image), renderingMode: .template) },
-                        centerView: { TitleListView(title: "Icon Place vs Background") },
+                        centerView: { TitleMoreDetailTextMessageListView(
+                            title: "Title",
+                            more: "More",
+                            detaile: "Detail",
+                            detaileMore: "More",
+                            subtitle: "Subtitle",
+                            tagText: "Percent",
+                            messageText: "Text message",
+                            infoImage: AdmiralUIResources.AssetSymbol.Service.Outline.info.image)
+                        },
                         trailingView: { ArrowListView() },
                         isSelected:
                             Binding(
                                 get: { self.selectedIndex == 3 },
                                 set: { _, _ in self.selectedIndex = self.selectedIndex == 3 ? nil : 3 }
                             ))
-                        .accessibilityElement()
-                        .accessibility(identifier: "ThirdImageTitleArrowListCell")
                         .disabled(isEnabledControlsState != 0)
+                        .accessibilityElement()
+                        .accessibility(identifier: "TitleMoreDetailTextMessageListView")
                     ListCell(
-                        leadingView: { ImageListView(image: Image(uiImage: Asset.Main.gem.image), renderingMode: .template) },
-                        centerView: { TitleListView(title: "Icon Place") },
+                        centerView: { TitleSubtitleButtonListView(
+                            title: "Title",
+                            tagSubtitle: "Subtitle",
+                            tagText: "Percent",
+                            subtitle: "Subtitle 2",
+                            buttonTitle: "Button",
+                            buttonAction: {})
+                        },
                         trailingView: { ArrowListView() },
                         isSelected:
                             Binding(
                                 get: { self.selectedIndex == 4 },
                                 set: { _, _ in self.selectedIndex = self.selectedIndex == 4 ? nil : 4 }
                             ))
-                        .accessibilityElement()
-                        .accessibility(identifier: "FourthImageTitleArrowListCell")
                         .disabled(isEnabledControlsState != 0)
+                        .accessibilityElement()
+                        .accessibility(identifier: "TitleSubtitleButtonListView")
                 }
                 .padding(.bottom, LayoutGrid.doubleModule * 4)
             }
