+++ ExampleiOS/Source/Classes/UserStrories/Main/SwiftUI/Cells/BaseCell/LeadingElements/LeadingElementsSwiftUIView.swift
@@ -1,17 +1,20 @@
 //
-//  LeadingComponentsSwiftUIView.swift
+//  ImageLableComponentsSwiftUIView.swift
 //  ExampleiOS
 //
-//  Created on 07.06.2021.
+//  Created on 22.06.2021.
 //
 
 import SwiftUI
 import AdmiralTheme
-import AdmiralUIResources
 import AdmiralSwiftUI
 
 @available(iOS 14.0.0, *)
-struct LeadingComponentsSwiftUIView: View {
+struct LeadingElementsSwiftUIView: View {
+    
+    enum Constants {
+        static let tabFormatAccessibilityValue = "%@. Page %i of %i"
+    }
     
     @State private var selectedIndex: Int?
     @State private var isEnabledControlsState: Int = 0
@@ -19,88 +22,77 @@ struct LeadingComponentsSwiftUIView: View {
     
     public var body: some View {
         let scheme = schemeProvider.scheme
-        NavigationContentView(navigationTitle: "Center elements") {
+        NavigationContentView(navigationTitle: "Leading elements") {
             scheme.backgroundColor.swiftUIColor
                 .edgesIgnoringSafeArea(.all)
-            ScrollView(showsIndicators: false) {
+            VStack(alignment: .leading, spacing: 0.0) {
                 StandardTab(items: [
                     StandartTabItem(text: "Default", accessibilityId: "SegmentControlDefault"),
                     StandartTabItem(text: "Disabled", accessibilityId: "SegmentControlDisabled")
                 ],
                             selection: $isEnabledControlsState)
                     .padding()
-                LazyVStack(alignment: .leading) {
+                ScrollView(showsIndicators: false) {
                     ListCell(
-                        centerView: { TitleListView(title: "Title") },
+                        leadingView: { LeadingCardListView(cardImage: Image(uiImage: Asset.Card.visa.image)) },
+                        centerView: { TitleListView(title: "Card Place") },
                         trailingView: { ArrowListView() },
                         isSelected:
                             Binding(
                                 get: { self.selectedIndex == 0 },
                                 set: { _, _ in self.selectedIndex = self.selectedIndex == 0 ? nil : 0 }
                             ))
-                        .disabled(isEnabledControlsState != 0)
                         .accessibilityElement()
-                        .accessibility(identifier: "TitleListView")
+                        .accessibility(identifier: "LeadingCardTitleArrowListCell")
+                        .disabled(isEnabledControlsState != 0)
                     ListCell(
-                        centerView: { TitleSubtitleListView(title: "Title", subtitle: "Subtitle") },
+                        leadingView: { ImageListView(image: Image(uiImage: Asset.Card.rnb.image)) },
+                        centerView: { TitleListView(title: "Lable Place") },
                         trailingView: { ArrowListView() },
                         isSelected:
                             Binding(
                                 get: { self.selectedIndex == 1 },
                                 set: { _, _ in self.selectedIndex = self.selectedIndex == 1 ? nil : 1 }
                             ))
-                        .disabled(isEnabledControlsState != 0)
                         .accessibilityElement()
-                        .accessibility(identifier: "TitleSubtitleListView")
+                        .accessibility(identifier: "FirstImageTitleArrowListCell")
+                        .disabled(isEnabledControlsState != 0)
                     ListCell(
-                        centerView: { SubtitleTitleListView(title: "Title", subtitle: "Subtitle") },
+                        leadingView: { ImageNameListView(text: "IN") },
+                        centerView: { TitleListView(title: "Icon Name") },
                         trailingView: { ArrowListView() },
                         isSelected:
                             Binding(
                                 get: { self.selectedIndex == 2 },
                                 set: { _, _ in self.selectedIndex = self.selectedIndex == 2 ? nil : 2 }
                             ))
-                        .disabled(isEnabledControlsState != 0)
                         .accessibilityElement()
-                        .accessibility(identifier: "SubtitleTitleListView")
+                        .accessibility(identifier: "SecondImageTitleArrowListCell")
+                        .disabled(isEnabledControlsState != 0)
                     ListCell(
-                        centerView: { TitleMoreDetailTextMessageListView(
-                            title: "Title",
-                            more: "More",
-                            detaile: "Detail",
-                            detaileMore: "More",
-                            subtitle: "Subtitle",
-                            tagText: "Percent",
-                            messageText: "Text message",
-                            infoImage: AdmiralUIResources.AssetSymbol.Service.Outline.info.image)
-                        },
+                        leadingView: { ImageBackgroundListView(image: Image(uiImage: Asset.Main.gem.image), renderingMode: .template) },
+                        centerView: { TitleListView(title: "Icon Place vs Background") },
                         trailingView: { ArrowListView() },
                         isSelected:
                             Binding(
                                 get: { self.selectedIndex == 3 },
                                 set: { _, _ in self.selectedIndex = self.selectedIndex == 3 ? nil : 3 }
                             ))
-                        .disabled(isEnabledControlsState != 0)
                         .accessibilityElement()
-                        .accessibility(identifier: "TitleMoreDetailTextMessageListView")
+                        .accessibility(identifier: "ThirdImageTitleArrowListCell")
+                        .disabled(isEnabledControlsState != 0)
                     ListCell(
-                        centerView: { TitleSubtitleButtonListView(
-                            title: "Title",
-                            tagSubtitle: "Subtitle",
-                            tagText: "Percent",
-                            subtitle: "Subtitle 2",
-                            buttonTitle: "Button",
-                            buttonAction: {})
-                        },
+                        leadingView: { ImageListView(image: Image(uiImage: Asset.Main.gem.image), renderingMode: .template) },
+                        centerView: { TitleListView(title: "Icon Place") },
                         trailingView: { ArrowListView() },
                         isSelected:
                             Binding(
                                 get: { self.selectedIndex == 4 },
                                 set: { _, _ in self.selectedIndex = self.selectedIndex == 4 ? nil : 4 }
                             ))
-                        .disabled(isEnabledControlsState != 0)
                         .accessibilityElement()
-                        .accessibility(identifier: "TitleSubtitleButtonListView")
+                        .accessibility(identifier: "FourthImageTitleArrowListCell")
+                        .disabled(isEnabledControlsState != 0)
                 }
                 .padding(.bottom, LayoutGrid.doubleModule * 4)
             }
